#21 Move test directories outside of src directories.
